add functions from assistant class to bot
