+++ funcs.py
@@ -2,6 +2,7 @@ import time
 import telebot
 import json
 
+
 def start_printed_timer(bot, message, time_in_seconds, chat_id):
     mes_id = bot.send_message(message.chat.id,
                               "Отсчет времени: {0}:{1}".format(time_in_seconds // 60 % 60, time_in_seconds % 60)).message_id
@@ -50,38 +51,6 @@ def get_task_by_name(name, tasks):
 		if task["name"] == name:
 			return [task_id, task]
 	return False
-
-def get_status_tomato(message, bot, task_id, time_start, time_tomato):
-	status = message.text
-	bot.send_message(message.from_user.id, "Теперь напиши комментарий")
-	bot.register_next_step_handler(message, lambda message: save_tomato(message, bot, task_id, time_start, status, time_tomato))
-
-def save_tomato(message, bot, task_id, time_start, status, time_tomato):
-	answer = message.text
-	user_id = str(message.from_user.id)
-	with open("test_data.json", "r", encoding="UTF-8") as file:
-		data = json.load(file)
-	if data[user_id]["tomatoes"] == {}:
-		tomato_id = "0"
-	else:
-		tomato_id = str(max(map(int, data[user_id]["tomatoes"].keys())) + 1)
-	data[user_id]["tomatoes"][tomato_id] = {'task_id': task_id, 'answer': answer, 'ts': time_start, 'status': status, 'time_tomato': time_tomato}
-	with open("test_data.json", "w", encoding="UTF-8") as file:
-		json.dump(data, file)
-	bot.send_message(message.from_user.id, "Не мне тебя судить, но я записал", reply_markup=get_keyboard_default())
-
-def register_task(bot, message, name, description):
-	user_id = str(message.from_user.id)
-	with open("test_data.json", "r", encoding="UTF-8") as file:
-		data = json.load(file)
-	if data[user_id]["tasks"] == {}:
-		task_id = "0"
-	else:
-		task_id = str(max(map(int, data[user_id]["tasks"].keys())) + 1)
-	data[user_id]["tasks"][task_id] = {"name": name, "description": description, "status": 0}
-	with open("test_data.json", "w", encoding="UTF-8") as file:
-		json.dump(data, file)
-	bot.send_message(message.from_user.id, "Готово, задача в списке доступных томатов", reply_markup=get_keyboard_default())
 
 def to_black_list(user_id, task_id):
 	with open("test_data.json", "r", encoding="UTF-8") as file:
