+++ main.py
@@ -2,8 +2,11 @@ import telebot
 import time
 import funcs
 from datetime import datetime
+import json
 
+from assistant import Assistant
 
+assistant = Assistant()
 tomato_time = 1
 
 bot = telebot.TeleBot('1660376392:AAHqeib5RhnCGZPUl8p-c4Fi8Cc5iB381PM')
@@ -14,6 +17,10 @@ keyboard_start_tomat.row('Начнем!')
 @bot.message_handler(commands=['start'])
 def start_message(message):
     funcs.register_id(str(message.from_user.id))
+    try:
+        assistant.add_user(str(message.from_user.id))
+    except ValueError:
+        print('Пользователь уже существует')
     bot.send_message(message.from_user.id, '''
     Привет! 
     Я бот, созданный помочь тебе устроить свое время так, чтобы ты смог работать максимально продуктивно. 
@@ -42,17 +49,52 @@ def register_task_name(message):
 
 def register_task_description(message, name):
     description = message.text
-    funcs.register_task(bot, message, name, description)
+    register_task(bot, message, name, description)
+
+def register_task(bot, message, name, description):
+    user_id = str(message.from_user.id)
+    with open("test_data.json", "r", encoding="UTF-8") as file:
+        data = json.load(file)
+    if data[user_id]["tasks"] == {}:
+        task_id = "0"
+    else:
+        task_id = str(max(map(int, data[user_id]["tasks"].keys())) + 1)
+    data[user_id]["tasks"][task_id] = {"name": name, "description": description, "status": 0}
+    with open("test_data.json", "w", encoding="UTF-8") as file:
+        json.dump(data, file)
+    assistant.users[str(message.from_user.id)].add_task(name, 6, description)
+    bot.send_message(message.from_user.id, "Готово, задача в списке доступных томатов", reply_markup=funcs.get_keyboard_default())
 
 def get_time_tomato(bot, message, task_id, task, chat_id):
     if message.text.lower() == 'стандартный томат':
         current_time = (datetime.now()).strftime("%m/%d/%Y, %H:%M:%S")
         funcs.start_printed_timer(bot, message, tomato_time * 60, chat_id)
         bot.send_message(message.from_user.id, task['description'])
-        bot.register_next_step_handler(message, lambda message: funcs.get_status_tomato(message, bot, task_id, current_time, tomato_time))
+        bot.register_next_step_handler(message, lambda message: get_status_tomato(message, bot, task_id, current_time, tomato_time))
     else:
         bot.send_message(message.from_user.id, 'Сколько минут длится томат?')
         bot.register_next_step_handler(message, lambda message: get_time_special_tomato(message, task_id, task, chat_id))
+
+def save_tomato(message, bot, task_id, time_start, status, time_tomato):
+    answer = message.text
+    user_id = str(message.from_user.id)
+    with open("test_data.json", "r", encoding="UTF-8") as file:
+        data = json.load(file)
+    if data[user_id]["tomatoes"] == {}:
+        tomato_id = "0"
+    else:
+        tomato_id = str(max(map(int, data[user_id]["tomatoes"].keys())) + 1)
+    data[user_id]["tomatoes"][tomato_id] = {'task_id': task_id, 'answer': answer, 'ts': time_start, 'status': status, 'time_tomato': time_tomato}
+    task_name = data[user_id]["tasks"][task_id]["name"]
+    with open("test_data.json", "w", encoding="UTF-8") as file:
+        json.dump(data, file)
+    assistant.users[str(message.from_user.id)].tasks[task_name].add_history(status, time_tomato, time_start, answer)
+    bot.send_message(message.from_user.id, "Не мне тебя судить, но я записал", reply_markup=funcs.get_keyboard_default())
+
+def get_status_tomato(message, bot, task_id, time_start, time_tomato):
+    status = message.text
+    bot.send_message(message.from_user.id, "Теперь напиши комментарий")
+    bot.register_next_step_handler(message, lambda message: save_tomato(message, bot, task_id, time_start, status, time_tomato))
 
 def get_time_special_tomato(message, task_id, task, chat_id):
     time_tomato = int(message.text)
